Add option to use closes rc file
For some projects, using a different rc file depending on the file being edited
is required. For example, consider a project whose tests require a less strict
set of rules than the rest of the code.

As an example, while we could care about line-around-directive in our production
code, we might not want to do so in our tests.

Setting `g:fixmyjs_use_rc_closest_to_file = 1` would make the plugin look for
the rc in each directory starting in the one containing the current file and up
to the project root directory. If no rc file is found, the previous logic goes
on.